NSA-7545 - Added last login filters
